countdown works but +3 hrs
